requirements.txt
Django==1.11.14
pytz
Markdown>=2.5.1
Pygments>=2.2.0
MySQL-python
beautifulsoup4>=4.3.2
Pillow
django-ckeditor===5.4.0